+++ examples/hello/src/error/mod.rs
@@ -2,7 +2,9 @@ use desirable::IntoResponse;
 use desirable::Response;
 use desirable::Result;
 use thiserror::Error;
+
 #[derive(Debug, Error)]
+#[expect(clippy::enum_variant_names, reason = "variant names read naturally at call sites")]
 pub enum Error {
   #[error("hyper error")]
   DesirableError(#[from] desirable::Error),